pyproject.toml
[project]
name = "dot-man"
version = "0.7.0"
description = "Dotfile manager with git-powered branching"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Bishoy Ehab" }
]
keywords = ["dotfiles", "git", "configuration", "cli"]
classifiers = [
    "Development Status :: 5 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.0",
    "cryptography>=41.0.0",
    "gitpython>=3.1",
    "rich>=13.0",
    "textual>=0.73.0",
    "tomli>=2.0; python_version < '3.11'",
    "tomlkit>=0.12.0",
    "questionary>=2.0.0",
]

[project.optional-dependencies]
tui = [
    "textual>=0.40",
]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "black>=24.8.0",
    "mypy>=1.14.1",
    "textual>=0.40",
    "ruff>=0.4.0",
]

[project.scripts]
dot-man = "dot_man.cli:main"

[project.urls]
Homepage = "https://github.com/BeshoyEhab/dot-man"
Repository = "https://github.com/BeshoyEhab/dot-man"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["dot_man*"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.mypy]
python_version = "3.10"
exclude = ["build/"]
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
